requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.36
pytesseract==0.3.13
Pillow==11.3.0
python-telegram-bot>=21.0
